NEW check_import_order.py: Initial commit of main parsing script
This is the intial commit for the main script in this repository. At
present, the script does what I need it to do on the OS/System I use,
which is Fedora 19 x86. My future intentions are to add selftests,
including running `pylint`, `pep8` and `nosetests` over this script,
and to check it against other Linux distros, as well as possibly OSX and
Windows (although the latter are unlikely unless someone else
contributes the tests and any fixes required).