NEW Makefile: `make clean` convenience function
This commit adds `Makefile` with a simple target to clean up any '*.pyc'
files. These are created when Python scripts are imported and are
'compiled' into byte-code. Noteably this happens when running the
`check_import_order.py` script with `nosetests`.

I intend to add a `make check` target which will run selftests for this
package. It will be useful to run `make clean` after `make check`.