check_import_order: add support for "from . import ..."
Previously, the regex for extracting the module name did not look for
the dot character. In situations where the import line was, for example,
"import module.submodule", the section of the import line *after* the
dot was ignored, whilst in situations where the import line was
"from . import ..." (i.e. from the local directory), an error was raised
because no module name was extracted.

This commit fixes the regex to accommodate dots in module titles including
when referencing the local directory. Note that if the module name is
detected as '.', we don't bother to import it in order to inspect it,
as we can infer immediately that it is a local import.

Selftests for the various situations have been added in parallel.