check_import_order: add comment regarding PYTHONPATH for local imports
This commit adds a comment in `construct_import_dict` to explain that
if we have not otherwise detected the module type and categorised thusly,
we assume that the import is valid and that it gets found on PYTHONPATH,
and assign it to the "local" category.

In the future I might make this check explicit and fail on any modules
which haven't otherwise been categorised.