README.rst: change reference from "system" to "standard" modules
In the initial commit of this repo I included the README which gave an
example of the command line usage for the script. The example made
reference to a "--system" command line option. When I actually added the
script, this had been changed to "--standard" (as in "standard library")
but I forgot to change the README as well.