check_import_order: [fix] get abspath of py_file to get dirname
Without getting the abspath of the py_file before getting the dirname of
the it, when the user passes a file which is located in the working
directory, e.g. check_import_order.py test.py, then the `local_dir` arg
which is passed to `construct_import_dict` would be an empty string
which would be detected as "not a directory" by `os.path.isdir`, and thus
the script would fail and exit fatally.